refactor: separate in checkbox
