+++ packages/components/checkbox/src/index.ts
@@ -1,14 +1,13 @@
-export { Checkbox, CheckboxIcon, useCheckbox } from "./checkbox"
+export { Checkbox, CheckboxIcon } from "./checkbox"
+export type { CheckboxIconProps, CheckboxProps } from "./checkbox"
+export { CheckboxGroup } from "./checkbox-group"
+export type { CheckboxGroupProps } from "./checkbox-group"
+export { useCheckbox } from "./use-checkbox"
+export type { UseCheckboxProps, UseCheckboxReturn } from "./use-checkbox"
+export { useCheckboxGroup } from "./use-checkbox-group"
+
 export type {
-  CheckboxIconProps,
-  CheckboxProps,
-  UseCheckboxProps,
-  UseCheckboxReturn,
-} from "./checkbox"
-export { CheckboxGroup, useCheckboxGroup } from "./checkbox-group"
-export type {
-  CheckboxGroupProps,
   CheckboxItem,
   UseCheckboxGroupProps,
   UseCheckboxGroupReturn,
-} from "./checkbox-group"
+} from "./use-checkbox-group"
